+++ core/src/Streamly/Internal/Data/Fold/Combinators.hs
@@ -1788,7 +1788,7 @@ distributeScan getFolds = Scanl consume initial extract final
 
     initial = return $ Partial (Tuple' [] [])
 
-    run st [] _ = return $ Partial st
+    run st [] _ = return $ Scanl.Partial st
     run (Tuple' ys zs) (Fold step init extr fin : xs) a = do
         res <- init
         case res of